feat: fincube #159 admin dashboard proxy integrations
