+++ frontend/src/pages/mfs/dashboard.tsx
@@ -14,10 +14,18 @@ import { ConnectButton } from "@rainbow-me/rainbowkit";
 import { useAccount, useConnect } from "wagmi";
 import { useEffect } from "react";
 import { toast, Toaster } from "sonner";
+import {
+  useLazyGetProposalCountQuery,
+  useLazyGetProposalThresholdQuery,
+} from "@redux/services/proxy";
 
 export default function MfsDashboard() {
   const { address } = useAccount();
   const { connectors } = useConnect();
+  const [getProposalThreshold] = useLazyGetProposalThresholdQuery();
+  const [getProposalCount] = useLazyGetProposalCountQuery();
+  // const [proposalCount, setProposalCount] = useState(0);
+  // const [proposalThreshold, setProposalThreshold] = useState(0);
 
   useEffect(() => {
     if (address) {
@@ -26,6 +34,33 @@ export default function MfsDashboard() {
       toast.error("Failed to connect wallet");
     }
   }, [address, connectors]);
+
+  useEffect(() => {
+    // getProposalCount
+    const fetchProposalCount = async () => {
+      try {
+        const response = await getProposalCount();
+        // setProposalCount(response);
+        console.log(response);
+      } catch (e) {
+        console.error(e);
+      }
+    };
+
+    // getProposalThreshold
+    const fetchProposalThreshold = async () => {
+      try {
+        const response = await getProposalThreshold();
+        // setProposalThreshold(response);
+        console.log(response);
+      } catch (e) {
+        console.error(e);
+      }
+    };
+
+    fetchProposalCount();
+    fetchProposalThreshold();
+  }, [getProposalCount, getProposalThreshold]);
 
   const auth = useSelector(
     (state: RootState) => state.persistedReducer.authReducer.auth
@@ -49,9 +84,8 @@ export default function MfsDashboard() {
       </div>
       <div className="flex flex-col gap-8 pt-28 px-8">
         <div className="flex gap-8 justify-center">
-          <div className="border rounded-xl p-8 flex flex-col items-center justify-center">
-            <h2 className="text-lg">Token Liquidity</h2>
-            <p className="text-2xl font-bold">0.7m USDT</p>
+          <div>
+            <DisplayCard title="Liquidity Ratio" value="0.7m USD" />
           </div>
           <div className="">
             <div className="w-64">
@@ -59,10 +93,10 @@ export default function MfsDashboard() {
             </div>
           </div>
           <div>
-            <DisplayCard title="Total Number of Customers" value={269} />
+            <DisplayCard title="Total Number of Proposals" value="10" />
           </div>
           <div>
-            <DisplayCard title="Annual Transfer Amount" value="1,000k BDT" />
+            <DisplayCard title="Proposal Threshold" value="10" />
           </div>
         </div>
