+++ frontend/src/redux/api/endpoints.ts
@@ -3,7 +3,7 @@ const BASE_URL_AUTH = BASE_URL + "/authentication";
 const BASE_URL_MFS = BASE_URL + "/mfs-business";
 const BASE_URL_DAO = BASE_URL + "/dao";
 const BASE_URL_PROPOSAL = BASE_URL + "/proposal-service";
-const BASE_URL_PROXY = BASE_URL + "/web3-proxy-service";
+const BASE_URL_PROXY = BASE_URL + "/web3-proxy-service/web3-dao-proxy";
 
 export const AUTH_ENDPOINTS = {
   BASE: BASE_URL_AUTH,
